Stop cloning the Settings inspect drawer; keep org value-plane fixtures on the provisioning view.
CatalogTypeDrawer shared the Settings inspect shell; extracting those sections into Settings (the owner until ticket 09) is the proper fix. EnvCells/Instances never belonged on the OC RT reconstruction — tests now inject them through a nil-in-production CatalogValuePlane.
